streaming-updater-producer: use long for sequence fields
While json does not actually make a distinction between long & int the
eventutilities Row -> Json serializer actually cared and failed with
convertion issues.
Force the use of a long type for sequence and sequence_length fields and
also improve tests so that they should capture such problems in the
future.

Bug: T374919

+++ streaming-updater-producer/src/main/scala/org/wikidata/query/rdf/updater/OutputEventStreamsBuilder.scala
@@ -204,8 +204,8 @@ class MutationDataChunkToRow(rowTypeInfo: EventRowTypeInfo, streamVersionMap: Ma
     output.setField("entity_id", value.getEntity)
     output.setField("rev_id", value.getRevision)
     output.setField("dt", value.getEventTime)
-    output.setField("sequence", value.getSequence)
-    output.setField("sequence_length", value.getSequenceLength)
+    output.setField("sequence", value.getSequence.toLong)
+    output.setField("sequence_length", value.getSequenceLength.toLong)
     output.setField("operation", value.getOperation)
 
     value match {
